Move Subscriptions model into h.models package
As part of an effort to move all models into the h.models package, this
moves the Subscriptions model out of h.notification and into h.models.
